redesigned docker build envirenment
It's still not that flexible of using a docker image as build
envirenment, as every time we want to modify the rootfs, a rebuild
is required. Inspired by lxc, now we put the entire rootfs outside
of the container, and just mount everything into it. As a result,
every modification is kept after the container is deleted, just a
real OS.

Signed-off-by: Ryan Wong <colorfulshark@gmail.com>