escalate permission with privileged container
It may cause some application to complain if rootfs files are not
set with correct ownership, while it's not elegant to ask for root
permission in the runtime. Here we do some trick where we gain root
privilege by operating in a privileged container. This should not be
the case as we need to create a privileged container anyway.

Signed-off-by: Ryan Wong <colorfulshark@gmail.com>